feat: Added GAAnimationBuilder
This class provides animations for GraphMolecule, simplifing the
animation process of subsections of the molecules.

It is created as an AnimationBuilder, a construct I just made up, which
uses the animation method of Mobjects to perform the desired actions
instead of inheriting from Animation. This was way easier even if it is
not completely pretty